+++ Endpoint/Intune/Windows/Compliance/Netskope/Detect-Netskope.ps1
@@ -6,10 +6,10 @@
 	Tags: intune endpoint MEM netskope
 
 .LICENSEURI
-https://github.com/bigjoestretch/powershell-toolkit/blob/main/LICENSE
+https://github.com/joelcottrell/powershell-toolkit/blob/main/LICENSE
 
 .PROJECTURI
-https://github.com/bigjoestretch/powershell-toolkit/tree/main/Endpoint/Intune/Windows/Compliance/Netskope
+https://github.com/joelcottrell/powershell-toolkit/tree/main/Endpoint/Intune/Windows/Compliance/Netskope
 
 .ICONURI