+++ Endpoint/Intune/Windows/Compliance/Zscaler/Detect-Zscaler.ps1
@@ -7,10 +7,10 @@
 	Tags: intune endpoint MEM zscaler
 
 .LICENSEURI
-https://github.com/bigjoestretch/powershell-toolkit/blob/main/LICENSE
+https://github.com/joelcottrell/powershell-toolkit/blob/main/LICENSE
 
 .PROJECTURI
-https://github.com/bigjoestretch/powershell-toolkit/tree/main/Endpoint/Intune/Windows/Compliance/Zscaler
+https://github.com/joelcottrell/powershell-toolkit/tree/main/Endpoint/Intune/Windows/Compliance/Zscaler
 
 .ICONURI