+++ Endpoint/Intune/Windows/ProactiveRemediations/Change-WinVerOEMInfo/Detect-WinVerOEMInfo.ps1
@@ -8,10 +8,10 @@
 	Tags: intune endpoint MEM winver oem
  
 .LICENSEURI
-https://github.com/bigjoestretch/powershell-toolkit/blob/main/LICENSE
+https://github.com/joelcottrell/powershell-toolkit/blob/main/LICENSE
 
 .PROJECTURI
-https://github.com/bigjoestretch/powershell-toolkit/tree/main/Endpoint/Intune/Windows/ProactiveRemediations/Change-WinVerOEMInfo
+https://github.com/joelcottrell/powershell-toolkit/tree/main/Endpoint/Intune/Windows/ProactiveRemediations/Change-WinVerOEMInfo
 
 .ICONURI