+++ Endpoint/Intune/Windows/ProactiveRemediations/Disable-RunCommand/Remediate-RunCommandRegistryKey.ps1
@@ -6,10 +6,10 @@
 	Tags: intune endpoint MEM proactive remediation disable run
 
 .LICENSEURI
-https://github.com/bigjoestretch/powershell-toolkit/blob/main/LICENSE
+https://github.com/joelcottrell/powershell-toolkit/blob/main/LICENSE
 
 .PROJECTURI
-https://github.com/bigjoestretch/powershell-toolkit/tree/main/Endpoint/Intune/Windows/ProactiveRemediations/Disable-RunCommand
+https://github.com/joelcottrell/powershell-toolkit/tree/main/Endpoint/Intune/Windows/ProactiveRemediations/Disable-RunCommand
 
 .ICONURI