+++ Endpoint/Intune/Windows/ProactiveRemediations/MicrosoftStore-ForcedAutoUpdate/Remediate-MicrosoftStoreAutoUpdate.ps1
@@ -6,10 +6,10 @@
 	Tags: intune endpoint MEM microsoft store
 
 .LICENSEURI
-https://github.com/bigjoestretch/powershell-toolkit/blob/main/LICENSE
+https://github.com/joelcottrell/powershell-toolkit/blob/main/LICENSE
 
 .PROJECTURI
-https://github.com/bigjoestretch/powershell-toolkit/tree/main/Endpoint/Intune/Windows/ProactiveRemediations/MicrosoftStore-ForcedAutoUpdate
+https://github.com/joelcottrell/powershell-toolkit/tree/main/Endpoint/Intune/Windows/ProactiveRemediations/MicrosoftStore-ForcedAutoUpdate
 
 .ICONURI