+++ Endpoint/Intune/Windows/ProactiveRemediations/WindowsUpdate/RemoveWUEntries/Remediate-WUPaths.ps1
@@ -7,10 +7,10 @@
 	Tags: intune endpoint MEM windows update
  
 .LICENSEURI
-https://github.com/bigjoestretch/powershell-toolkit/blob/main/LICENSE
+https://github.com/joelcottrell/powershell-toolkit/blob/main/LICENSE
 
 .PROJECTURI
-https://github.com/bigjoestretch/powershell-toolkit/tree/main/Endpoint/Intune/Windows/ProactiveRemediations/WindowsUpdate/RemoveWUEntries
+https://github.com/joelcottrell/powershell-toolkit/tree/main/Endpoint/Intune/Windows/ProactiveRemediations/WindowsUpdate/RemoveWUEntries
 
 .ICONURI