+++ Endpoint/Intune/Windows/Win32Apps/SnagIT-2024/Uninstall-Snagit2024.ps1
@@ -6,10 +6,10 @@
 	Tags: intune endpoint MEM snagit 2024
 
 .LICENSEURI
-https://github.com/bigjoestretch/powershell-toolkit/blob/main/LICENSE
+https://github.com/joelcottrell/powershell-toolkit/blob/main/LICENSE
 
 .PROJECTURI
-https://github.com/bigjoestretch/powershell-toolkit/tree/main/Endpoint/Intune/Windows/Win32Apps/SnagIT-2024
+https://github.com/joelcottrell/powershell-toolkit/tree/main/Endpoint/Intune/Windows/Win32Apps/SnagIT-2024
 
 .ICONURI