+++ Windows/Common/Netsh/Start-NetshMultiTrace.ps1
@@ -6,10 +6,10 @@
 	Tags: windows netsh
 
 .LICENSEURI
-https://github.com/bigjoestretch/powershell-toolkit/blob/main/LICENSE
+https://github.com/joelcottrell/powershell-toolkit/blob/main/LICENSE
 
 .PROJECTURI
-https://github.com/bigjoestretch/powershell-toolkit/tree/main/Windows/Common/Netsh
+https://github.com/joelcottrell/powershell-toolkit/tree/main/Windows/Common/Netsh
 
 .ICONURI